#22 added a new way of isomorphism check, unifying CSP and combinatorial
the new variant kind of is the best of the two worlds of both the
CSP-based and the combinatorial approach. it doesn't use actual trees,
but it kind of simulates them - it drastically decreases the number of
combinations to check and the number of the checks themselves. it does
support graphs that are not connected and it runs correct and much
faster in all tests. so the combinatorial and CSP-based approaches
should be completely replaced by this new approach.