#36 trying to do something with stochastic regular expressions
doesn't seem right though. this expression library ain't made for this
either (gotta parse probabilities out of edge labels -> can't even parse
to number and calculate within the expression library; could just
extract the probability as a string or calculate with numbers -> no way
to extract a string and parse it into a number without changing the
expression library).

on the other hand the paper doesn't describe quite right, how the
probabilities are calculated, and it's somehow open to interpretation.
as far as i understand you can always look at everything except star as
an unary or binary concatenation and put a constant rate as probability
to it. and with star you can just take the old value. so it then would
be possible (but) pointless, to do it with just string operations. i was
trying to do this. but the expression library keeps throwing some
errors. i don't know if it's worth the hastle